refactor and type changes
