+++ app/react/V2/Routes/Entity/Components/index.ts
@@ -1,7 +1,6 @@
-export * from './urlParams';
+export * from '../urlParams';
 export { PDFView } from './PDFView';
 export { TabLabel } from './TabLabel';
 export { SearchResults } from './SearchResults';
 export { SearchHintsModal } from './SearchHintsModal';
 export { ToCPanel } from './ToCPanel';
-export type { LoaderResponse } from './types';
