+++ app/react/V2/Routes/Entity/index.ts
@@ -1,0 +1,2 @@
+export { entityLoader } from './loader';
+export { Entity } from './Entity';
