+++ app/react/V2/Routes/Entity/loader.ts
@@ -1,0 +1,113 @@
+import { IncomingHttpHeaders } from 'http';
+import { FetchResponseError } from 'shared/JSONRequest';
+import { LoaderFunction } from 'react-router';
+import { getPagePlaintext } from 'V2/api/files';
+import { snippets } from 'V2/api/search';
+import { SnippetsSearchResponse } from 'V2/api/types';
+import { getEntityCompositionUseCase } from 'V2/application/container/singletons';
+import { fullDetailOptions } from 'V2/application/optionsPresets';
+import { entityLoaderCache } from './EntityLoaderCache';
+import { PAGE_PARAM, SEARCH_PARAM } from './Components';
+import { LoaderResponse } from './types';
+
+const entityLoader =
+  (headers?: IncomingHttpHeaders): LoaderFunction =>
+  // eslint-disable-next-line max-statements
+  async ({ params, request }): Promise<LoaderResponse> => {
+    const entitySharedId = params.sharedId;
+    const { searchParams } = new URL(request.url);
+    const currentPage = searchParams.get(PAGE_PARAM) || '1';
+    const currentSearchTerm = searchParams.get(SEARCH_PARAM);
+
+    if (!entitySharedId) {
+      return undefined;
+    }
+
+    let entity = entityLoaderCache.getEntity(entitySharedId, 'en');
+    let pagePlaintext: string | undefined = '';
+    let searchResults: SnippetsSearchResponse | undefined;
+
+    if (!entity) {
+      const entityCompositionUseCase = await getEntityCompositionUseCase();
+
+      const composition = await entityCompositionUseCase.composeEntity(
+        entitySharedId,
+        fullDetailOptions,
+        {
+          headers,
+        }
+      );
+
+      if (!composition.success || !composition.entity) {
+        throw new Response(
+          JSON.stringify({
+            error: 'Failed to load entity',
+            message: composition.error || 'Entity not found',
+            entityId: entitySharedId,
+          }),
+          {
+            status: 404,
+            statusText: 'Entity Not Found',
+            headers: {
+              'Content-Type': 'application/json',
+            },
+          }
+        );
+      }
+
+      entity = composition.entity;
+      entityLoaderCache.setEntity(entitySharedId, 'en', entity);
+
+      if (composition.entity.mainDocument?._id) {
+        pagePlaintext = entityLoaderCache.getPlaintext(
+          composition.entity.mainDocument._id?.toString(),
+          Number(currentPage)
+        );
+
+        if (pagePlaintext) {
+          const response = await getPagePlaintext(
+            composition.entity.mainDocument?._id as string,
+            Number.parseInt(currentPage, 10)
+          );
+
+          if (response instanceof FetchResponseError) {
+            throw new Response(
+              JSON.stringify({
+                error: 'Failed to load plaintext',
+                message: response.message,
+                entityId: entitySharedId,
+              }),
+              {
+                status: 404,
+                statusText: 'Failed to load plaintext',
+                headers: {
+                  'Content-Type': 'application/json',
+                },
+              }
+            );
+          } else {
+            pagePlaintext = response;
+            entityLoaderCache.setPlaintext(entitySharedId, Number(currentPage), pagePlaintext);
+          }
+        }
+      }
+    }
+
+    if (currentSearchTerm) {
+      searchResults = entityLoaderCache.getSearchResults(entitySharedId, currentSearchTerm);
+
+      if (!searchResults) {
+        searchResults = await snippets({
+          sharedId: entity.sharedId,
+          limit: 0,
+          searchString: currentSearchTerm,
+        });
+
+        entityLoaderCache.setSearchResults(entitySharedId, currentSearchTerm, searchResults);
+      }
+    }
+
+    return { entity, pagePlaintext, searchResults };
+  };
+
+export { entityLoader };
